# fix(p2p): Avoid generating p2p keys in DefaultP2P, for faster tests